Made CategoryBuilder a top-level class
